# Use json-iterator instead of ugorji for JSON.