--- a/pkg/controller/garbagecollector/metaonly/BUILD
+++ b/pkg/controller/garbagecollector/metaonly/BUILD
@@ -10,19 +10,16 @@ go_library(
     name = "go_default_library",
     srcs = [
         "metaonly.go",
-        "types.generated.go",
         "types.go",
         "zz_generated.deepcopy.go",
     ],
     deps = [
         "//pkg/api:go_default_library",
-        "//vendor/github.com/ugorji/go/codec:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/serializer:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
     ],
 )