--- a/pkg/kubectl/testing/BUILD
+++ b/pkg/kubectl/testing/BUILD
@@ -1,25 +1,17 @@
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
     srcs = [
         "doc.go",
-        "types.generated.go",
         "types.go",
         "zz_generated.deepcopy.go",
     ],
-    visibility = [
-        "//build/visible_to:pkg_kubectl_testing_CONSUMERS",
-    ],
+    visibility = ["//visibility:public"],
     deps = [
-        "//vendor/github.com/ugorji/go/codec:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
     ],
 )
 
@@ -27,13 +19,12 @@ filegroup(
     name = "package-srcs",
     srcs = glob(["**"]),
     tags = ["automanaged"],
+    visibility = ["//visibility:private"],
 )
 
 filegroup(
     name = "all-srcs",
     srcs = [":package-srcs"],
     tags = ["automanaged"],
-    visibility = [
-        "//build/visible_to:pkg_kubectl_testing_CONSUMERS",
-    ],
+    visibility = ["//visibility:public"],
 )