--- a/staging/src/k8s.io/api/policy/v1beta1/BUILD
+++ b/staging/src/k8s.io/api/policy/v1beta1/BUILD
@@ -11,7 +11,6 @@ go_library(
         "doc.go",
         "generated.pb.go",
         "register.go",
-        "types.generated.go",
         "types.go",
         "types_swagger_doc_generated.go",
         "zz_generated.deepcopy.go",
@@ -19,12 +18,10 @@ go_library(
     deps = [
         "//vendor/github.com/gogo/protobuf/proto:go_default_library",
         "//vendor/github.com/gogo/protobuf/sortkeys:go_default_library",
-        "//vendor/github.com/ugorji/go/codec:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/conversion:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
     ],
 )