… Godeps/Godeps.json
… federation/apis/federation/v1beta1/types.generated.go
… hack/update-codecgen.sh
… hack/verify-codecgen.sh
… pkg/api/serialization_test.go
… pkg/controller/garbagecollector/metaonly/types.generated.go
… pkg/kubectl/testing/types.generated.go
… staging/src/k8s.io/api/Godeps/Godeps.json
… staging/src/k8s.io/api/admission/v1alpha1/types.generated.go
… staging/src/k8s.io/api/admissionregistration/v1alpha1/types.generated.go
… staging/src/k8s.io/api/apps/v1beta1/types.generated.go
… staging/src/k8s.io/api/authentication/v1/types.generated.go
… staging/src/k8s.io/api/authentication/v1beta1/types.generated.go
… staging/src/k8s.io/api/authorization/v1/types.generated.go
… staging/src/k8s.io/api/authorization/v1beta1/types.generated.go
… staging/src/k8s.io/api/autoscaling/v1/types.generated.go
… staging/src/k8s.io/api/autoscaling/v2alpha1/types.generated.go
… staging/src/k8s.io/api/batch/v1/types.generated.go
… staging/src/k8s.io/api/batch/v1beta1/types.generated.go
… staging/src/k8s.io/api/batch/v2alpha1/types.generated.go
… staging/src/k8s.io/api/certificates/v1beta1/types.generated.go
… staging/src/k8s.io/api/core/v1/types.generated.go
… staging/src/k8s.io/api/extensions/v1beta1/types.generated.go
… staging/src/k8s.io/api/imagepolicy/v1alpha1/types.generated.go
… staging/src/k8s.io/api/networking/v1/BUILD
… staging/src/k8s.io/api/networking/v1/types.generated.go
… staging/src/k8s.io/api/policy/v1beta1/types.generated.go
… staging/src/k8s.io/api/rbac/v1/types.generated.go
… staging/src/k8s.io/api/rbac/v1alpha1/types.generated.go
… staging/src/k8s.io/api/rbac/v1beta1/types.generated.go