@thockin 2017-07-02
This has hacked up godeps, but other than that - it actually passes tests now!  Do we have a relevant benchmark?  @smarterclayton, you had other concerns?