@thockin 2017-07-02
FWIW by the micro benchmarks, it shows Marshal 2x faster than stdlib and Unmarshal 3x or more faster.