@smarterclayton 2017-07-03
Other than custom marshallers, the registration for custom marshallers
(which assumed package short name + type, which is not unique enough), and
the possibility that the unsafe code is really unsafe, no.

On Sun, Jul 2, 2017 at 1:18 AM, k8s-ci-robot <notifications@github.com>
wrote:

> @thockin <https://github.com/thockin>: The following tests *failed*, say
> /retest to rerun them all:
> Test name Commit Details Rerun command
> pull-kubernetes-bazel 5e9cef0
> <https://github.com/kubernetes/kubernetes/commit/5e9cef0cbf1b3e0e6ea0748f00330209de8b7585>
> link
> <https://k8s-gubernator.appspot.com/build/kubernetes-jenkins/pr-logs/pull/48287/pull-kubernetes-bazel/28460/> /test
> pull-kubernetes-bazel
> pull-kubernetes-cross ecb27da
> <https://github.com/kubernetes/kubernetes/commit/ecb27da5964f493bfdf5e38f71b50c04a2ce17ba>
> link
> <https://k8s-gubernator.appspot.com/build/kubernetes-jenkins/pr-logs/pull/48287/pull-kubernetes-cross/3131/> /test
> pull-kubernetes-cross
> pull-kubernetes-verify ecb27da
> <https://github.com/kubernetes/kubernetes/commit/ecb27da5964f493bfdf5e38f71b50c04a2ce17ba>
> link
> <https://k8s-gubernator.appspot.com/build/kubernetes-jenkins/pr-logs/pull/48287/pull-kubernetes-verify/37603/> /test
> pull-kubernetes-verify
>
> Full PR test history <https://k8s-gubernator.appspot.com/pr/48287>. Your
> PR dashboard <https://k8s-gubernator.appspot.com/pr/thockin>. Please help
> us cut down on flakes by linking to
> <https://github.com/kubernetes/community/blob/master/contributors/devel/flaky-tests.md#filing-issues-for-flaky-tests>
> an open issue
> <https://github.com/kubernetes/kubernetes/issues?q=is:issue+is:open> when
> you hit one in your PR.
>
> Instructions for interacting with me using PR comments are available here
> <https://github.com/kubernetes/community/blob/master/contributors/devel/pull-requests.md>.
> If you have questions or suggestions related to my behavior, please file an
> issue against the kubernetes/test-infra
> <https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:>
> repository. I understand the commands that are listed here
> <https://github.com/kubernetes/test-infra/blob/master/commands.md>.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/48287#issuecomment-312471594>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p6bW6xN9gq6WRg3to3W5yYyqV9Nyks5sJyg-gaJpZM4OJxm7>
> .
>