@thockin 2017-07-04
More fuzzing is more better.  I added a ton of tests to the json-iterator repo to fuzz just about every combination of types and structs and tags I could think up.  Many a direct result of making these tests pass.  We have the serialization test already - can we expand on that?  Do the same thing, but for our own complex types?

Or we could fuzz at the input level - generate random structs, write to files, run the round-trip test against them.