@smarterclayton 2017-07-04
I'm wondering if we should be JSON fuzzing (since this is a net new custom
parser).  Not just well formed input.

On Tue, Jul 4, 2017 at 2:18 PM, Tim Hockin <notifications@github.com> wrote:

> More fuzzing is more better. I added a ton of tests to the json-iterator
> repo to fuzz just about every combination of types and structs and tags I
> could think up. Many a direct result of making these tests pass. We have
> the serialization test already - can we expand on that? Do the same thing,
> but for our own complex types?
>
> Or we could fuzz at the input level - generate random structs, write to
> files, run the round-trip test against them.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/48287#issuecomment-312930888>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p9P2r4OeaL8BaLWK1Tx8eIWsQsu8ks5sKoHogaJpZM4OJxm7>
> .
>