@thockin 2017-07-17
I added random-input fuzzing to json-iterator which, no surprise, caught some bugs.  They are working on those bugs, mostly fixed, one last one to do it sounds like.

https://github.com/json-iterator/go/blob/master/unmarshal_input_test.go