@smarterclayton 2017-07-17
Alignment bugs or unsafe memory access bugs?

On Mon, Jul 17, 2017 at 11:54 AM, Tim Hockin <notifications@github.com>
wrote:

> I added random-input fuzzing to json-iterator which, no surprise, caught
> some bugs. They are working on those bugs, mostly fixed, one last one to do
> it sounds like.
>
> https://github.com/json-iterator/go/blob/master/unmarshal_input_test.go
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/48287#issuecomment-315784043>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABG_p6D1tVKuQehYpZF9d4zRtl96zBcFks5sO4JXgaJpZM4OJxm7>
> .
>