@thockin 2017-07-18
Just parsing bugs - allowed invalid input when it shouldn't have.  To test alignment would require deeper understanding of their internals.